VAD Completed
Completed Calculation of Voice Activity Detection based on STE and ZCR,
In their linear combination, accurate results were yielded only if the
scaling factor was one (and not more). Check newly commited images for the
box plot. Now one thing that remains  is about how to plot the frames or
segments of the audio signal (hilbert? or manual calculations?)