Complete implementation of 10 Digit's HMMs
This commit ensured that yahmm support multi-dimensional coefficients
and can support infinite HMM generators, which is just good for
applying cepstral coefficients in order to test and train

Further code will include some file handling stuffs, and some
statistical tests need to be written.

For demo purposes, the trained HMMs will be written out to file
stream and then later used in other programs. (Just for demo)
All the results and discussions will be based on the accuracy of
implementaion and test results,

If possible, we will aim to create a null hypothesis and calculate
the significance of the result, so that we could publish it in the
future if need be.

Signing off,
@NootanGhimire